add checks for file exists and image file format

+++ pub/sandbox/index.php
@@ -21,14 +21,30 @@
     {
         //zdefiniuj folder do którego trafią pliki (ścieżka względem pliku index.php)
         $targetDir = "img/";
+
         //pobierz pierwotną nazwę pliku z tablicy $_FILES
         $sourceFileName = $_FILES['uploadedFile']['name'];
+
         //pobierz tymczasową ścieżkę do pliku na serwerze
         $tempURL = $_FILES['uploadedFile']['tmp_name'];
+
+        //sprawdź czy mamy do czynienia z obrazem
+        $imgInfo = getimagesize($tempURL);
+        if(!is_array($imgInfo)) {
+            die("BŁĄD: Przekazany plik nie jest obrazem!");
+        }
+
         //zbuduj docelowy URL pliku na serwerze
         $targetURL = $targetDir . $sourceFileName;
+
+        //sprawdź czy plik przypadkiem już nie istnieje
+        if(file_exists($targetURL)) {
+            die("BŁĄD: Podany plik już istnieje!");
+        }
+
         //przesuń plik do docelowej lokalizacji
         move_uploaded_file($tempURL, $targetURL);
+        echo "Plik został poprawnie wgrany na serwer";
     }
     ?>
 </body>
